+++ packages/react/src/lib/components/Furniture/Bands/HorizontalBand/HorizontalBand.unit.tsx
@@ -1,4 +1,4 @@
-import * as d3 from "@d3-chart/d3";
+import * as d3 from "@chart-it/d3";
 import { Provider } from "react-redux";
 import React from "react";
 import { render } from "@testing-library/react";