+++ packages/react/src/lib/components/Shapes/Line/Line.tsx
@@ -1,4 +1,4 @@
-import type { IColor } from "@d3-chart/types";
+import type { IColor } from "@chart-it/types";
 import React from "react";
 
 export interface ILineProps {
@@ -19,5 +19,5 @@ export function Line({ fill }: ILineProps) {
         flexShrink: 0,
     };
 
-    return <div className="d3-chart indicator-line" style={style} />;
+    return <div className="chart-it indicator-line" style={style} />;
 }