+++ packages/react/src/lib/components/Shapes/Square/Square.tsx
@@ -1,4 +1,4 @@
-import type { IColor } from "@d3-chart/types";
+import type { IColor } from "@chart-it/types";
 import React from "react";
 
 export interface ISquareProps {
@@ -19,5 +19,5 @@ export function Square({ fill }: ISquareProps) {
         flexShrink: 0,
     };
 
-    return <div className="d3-chart indicator-square" style={style} />;
+    return <div className="chart-it indicator-square" style={style} />;
 }