+++ packages/react/src/lib/components/TooltipOverlay/Tooltip/TooltipItem/TooltipItem.tsx
@@ -1,4 +1,4 @@
-import type { ITooltipFormatter, ITooltipItem } from "@d3-chart/types";
+import type { ITooltipFormatter, ITooltipItem } from "@chart-it/types";
 import React from "react";
 import { useSelector } from "react-redux";
 
@@ -49,13 +49,13 @@ export function TooltipItem({ name, value, icon, fill, prefix, suffix, format = 
     };
 
     return (
-        <div className="d3-chart tooltip-item" style={styles.tooltipItem}>
+        <div className="chart-it tooltip-item" style={styles.tooltipItem}>
             {Shape && <Shape fill={fill} />}
-            <div className="d3-chart tooltip-values" style={styles.tooltipValues}>
-                <span className="d3-chart tooltip-series-name" style={styles.tooltipSeriesName}>
+            <div className="chart-it tooltip-values" style={styles.tooltipValues}>
+                <span className="chart-it tooltip-series-name" style={styles.tooltipSeriesName}>
                     {name}:
                 </span>
-                <span className="d3-chart tooltip-series-value">{formattedValue}</span>
+                <span className="chart-it tooltip-series-value">{formattedValue}</span>
             </div>
         </div>
     );