+++ packages/react/src/lib/hoc/canvas/virtual/renderVirtualElements.ts
@@ -1,5 +1,5 @@
-import * as d3 from "@d3-chart/d3";
-import type { Transition } from "@d3-chart/d3";
+import * as d3 from "@chart-it/d3";
+import type { Transition } from "@chart-it/d3";
 
 import { renderElements } from "../renderElements";