+++ packages/react/src/lib/store/chart/tests/chartActions.unit.ts
@@ -1,4 +1,4 @@
-import * as d3 from "@d3-chart/d3";
+import * as d3 from "@chart-it/d3";
 
 import { chartActions } from "../chartActions";
 import { themes } from "../../../themes";