+++ packages/react/src/lib/utils/checks/ensureNoScaleOverflow.ts
@@ -1,5 +1,5 @@
-import * as d3 from "@d3-chart/d3";
-import type { IData, IScale } from "@d3-chart/types";
+import * as d3 from "@chart-it/d3";
+import type { IData, IScale } from "@chart-it/types";
 
 /**
  * Validates that the scale can cope with aggregated data